Actualización clase Persona
- Se cambió dni por long
- Se eliminó la variable edad
- Se agregó la variable fechaNac y Apellido
- Actualización en la bdd

+++ src/main/java/com/owo/TP_prg3/Clases/Cliente/modelo/ClienteRepositorio.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface ClienteRepositorio extends JpaRepository<Cliente,Long> {
 
-    Optional<Cliente> findByPersona_Dni(int dni);
+    Optional<Cliente> findByPersona_Dni(Long dni);
 }
