+++ src/main/java/com/owo/TP_prg3/Clases/Empleado/modelo/EmpleadoRepositorio.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface EmpleadoRepositorio extends JpaRepository<Empleado, Long> {
 
-    Optional<Empleado> findByPersona_Dni(int dni);
+    Optional<Empleado> findByPersona_Dni(Long dni);
 
     Optional<Empleado> findByUsuario_Email(String trim);
