+++ src/main/java/com/owo/TP_prg3/Clases/Persona/modelo/PersonaRepositorio.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 
 @Repository
 public interface PersonaRepositorio extends JpaRepository<Persona, Long> {
-    Optional<Persona> findByDni(int dni);
+    Optional<Persona> findByDni(Long dni);
 }
